Publish Code for Storing customer information
Project to store the customer’s information in a secure database. The database is locked with a login username and password. One can store the customer’s first and last name, address, phone-number and email address into the database created. The ability to store the details would be through a graphical user interface coded in Java.